Extract inner classes to seperate files
