+++ core/src/main/java/com/novoda/accessibility/CaptionManager.java
@@ -1,0 +1,7 @@
+package com.novoda.accessibility;
+
+interface CaptionManager {
+
+    boolean isClosedCaptioningEnabled();
+
+}
