+++ core/src/main/java/com/novoda/accessibility/DummyClosedCaptionManager.java
@@ -1,0 +1,14 @@
+package com.novoda.accessibility;
+
+class DummyClosedCaptionManager implements CaptionManager {
+
+    DummyClosedCaptionManager() {
+        // no-op
+    }
+
+    @Override
+    public boolean isClosedCaptioningEnabled() {
+        return false;
+    }
+
+}
